+++ accounting/src/sdk/models/shared/companydataset.ts
@@ -65,7 +65,7 @@ export class Phone extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     @Expose({ name: "number" })
-    number: string;
+    number?: string;
 
     /**
      * The type of phone number