+++ accounting/src/sdk/models/shared/phonenumberitems.ts
@@ -12,7 +12,7 @@ export class PhoneNumberItems extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     @Expose({ name: "number" })
-    number: string;
+    number?: string;
 
     /**
      * The type of phone number