+++ accounting/src/sdk/models/shared/purchaseorder.ts
@@ -11,6 +11,32 @@ import { SupplierRef } from "./supplierref";
 import { Expose, Type } from "class-transformer";
 
 /**
+ * The user who created the purchase order in the accounting system
+ */
+export class User extends SpeakeasyBase {
+    /**
+     * Email address of the user.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "email" })
+    email?: string;
+
+    /**
+     * First name of the user.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "firstName" })
+    firstName?: string;
+
+    /**
+     * Last name of the user.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "lastName" })
+    lastName?: string;
+}
+
+/**
  * > View the coverage for purchase orders in the <a className="external" href="https://knowledge.codat.io/supported-features/accounting?view=tab-by-data-type&dataType=purchaseOrders" target="_blank">Data coverage explorer</a>.
  *
  * @remarks
@@ -22,6 +48,14 @@ import { Expose, Type } from "class-transformer";
  * This information can be used to provide visibility on a business's expected payables and to track a purchase through the full procurement process.
  */
 export class PurchaseOrder extends SpeakeasyBase {
+    /**
+     * The user who created the purchase order in the accounting system
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "createdBy" })
+    @Type(() => User)
+    createdBy?: User;
+
     /**
      * The currency data type in Codat is the [ISO 4217](https://en.wikipedia.org/wiki/ISO_4217) currency code, e.g. _GBP_.
      *